Test: Extract function to run rpmbuild
This also more clearly prints what it's doing so that it's easier to
see exactly what the output of any particular call to `rpmbuild`
printed. This is helpful because `rpmbuild` will run all stages to a
particular point; we request stages with separate calls to rpmbuild
causing them to be run again.