Test: Finish ct1 package build and display contents
It turns out that the final package build removes the work dir under
BUILDROOT created by the %install step; thus we need to use `-bi` when
trying to inspect intermediate output. However, `-bi` appears to do
all the various checks to ensure that the final RPM can be properly
created.